Add two column layout choropleth tile topical page
