+++ packages/app/src/pages/actueel/veiligheidsregio/[code].tsx
@@ -31,6 +31,7 @@ import { createPositiveTestedPeopleRegionalTooltip } from '~/components/chorople
 import { escalationTooltip } from '~/components/choropleth/tooltips/region/escalation-tooltip';
 import { FCWithLayout, getDefaultLayout } from '~/domain/layout/layout';
 import { ArticleList } from '~/domain/topical/article-list';
+import { ChoroplethTwoColumnLayout } from '~/domain/topical/choropleth-two-column-layout';
 import { DataSitemap } from '~/domain/topical/data-sitemap';
 import { EditorialSummary } from '~/domain/topical/editorial-teaser';
 import { EditorialTile } from '~/domain/topical/editorial-tile';
@@ -38,7 +39,6 @@ import { EscalationLevelExplanations } from '~/domain/topical/escalation-level-e
 import { MiniTrendTile } from '~/domain/topical/mini-trend-tile';
 import { MiniTrendTileLayout } from '~/domain/topical/mini-trend-tile-layout';
 import { TopicalChoroplethContainer } from '~/domain/topical/topical-choropleth-container';
-import { TopicalChoroplethTile } from '~/domain/topical/topical-choropleth-layout';
 import { TopicalSectionHeader } from '~/domain/topical/topical-section-header';
 import { TopicalTile } from '~/domain/topical/topical-tile';
 import { topicalPageQuery } from '~/queries/topical-page-query';
@@ -276,12 +276,9 @@ const TopicalSafetyRegion: FCWithLayout<typeof getStaticProps> = (props) => {
               </TopicalTile>
             </Box>
 
-            <TopicalChoroplethTile
-              title={
-                siteText.common_actueel.secties.positief_getest_kaart.titel
-              }
-              choropleth={
-                <>
+            <TopicalTile>
+              <ChoroplethTwoColumnLayout>
+                <Box>
                   {selectedMap === 'municipal' && (
                     <MunicipalityChoropleth
                       data={choropleth.gm}
@@ -320,25 +317,32 @@ const TopicalSafetyRegion: FCWithLayout<typeof getStaticProps> = (props) => {
                       )}
                     />
                   )}
-                </>
-              }
-            >
-              <Text css={css({ maxWidth: 350 })}>
-                {siteText.positief_geteste_personen.map_toelichting}
-              </Text>
-              <Box mb={5}>
-                <ChartRegionControls
-                  value={selectedMap}
-                  onChange={setSelectedMap}
-                />
-              </Box>
-              <ChoroplethLegenda
-                thresholds={regionThresholds.tested_overall.infected_per_100k}
-                title={
-                  siteText.positief_geteste_personen.chloropleth_legenda.titel
-                }
-              />
-            </TopicalChoroplethTile>
+                </Box>
+                <Box>
+                  <Text css={css({ maxWidth: 350 })}>
+                    {siteText.positief_geteste_personen.map_toelichting}
+                  </Text>
+                  <Box
+                    mb={4}
+                    css={css({ '> div': { justifyContent: 'flex-start' } })}
+                  >
+                    <ChartRegionControls
+                      value={selectedMap}
+                      onChange={setSelectedMap}
+                    />
+                  </Box>
+                  <ChoroplethLegenda
+                    thresholds={
+                      regionThresholds.tested_overall.infected_per_100k
+                    }
+                    title={
+                      siteText.positief_geteste_personen.chloropleth_legenda
+                        .titel
+                    }
+                  />
+                </Box>
+              </ChoroplethTwoColumnLayout>
+            </TopicalTile>
 
             <DataSitemap />
